# src/main/resources/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Todo API - Javalin",
    "version": "1.0.0",
    "description": "API for managing persons, items and tasks."
  },
  "paths": {
    "/person": {
      "get": {
        "summary": "Get all persons",
        "responses": {
          "200": {
            "description": "List persons",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/PersonResponse" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new person",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PersonCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Person created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PersonResponse" }
              }
            }
          }
        }
      }
    },

    "/person/{id}": {
      "get": {
        "summary": "Get a person by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Person found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PersonResponse" }
              }
            }
          },
          "404": { "description": "Person not found" }
        }
      },
      "patch": {
        "summary": "Update a person partially",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PersonUpdate" }
            }
          }
        },
        "responses": {
          "200": { "description": "Person updated" }
        }
      },
      "delete": {
        "summary": "Delete a person",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "Deleted successfully" }
        }
      }
    },

    "/items": {
      "get": {
        "summary": "List all items",
        "responses": {
          "200": {
            "description": "List items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TodoItemResponse" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TodoItemCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoItemResponse" }
              }
            }
          }
        }
      }
    },

    "/items/{id}": {
      "get": {
        "summary": "Get an item by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Item found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoItemResponse" }
              }
            }
          },
          "404": { "description": "Item not found" }
        }
      },
      "patch": {
        "summary": "Update an item partially",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TodoItemUpdate" }
            }
          }
        },
        "responses": {
          "200": { "description": "Item updated" }
        }
      },
      "delete": {
        "summary": "Delete an item",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "Deleted successfully" }
        }
      }
    },

    "/tasks": {
      "get": {
        "summary": "List all tasks",
        "responses": {
          "200": {
            "description": "List tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TodoTaskResponse" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TodoTaskCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoTaskResponse" }
              }
            }
          }
        }
      }
    },

    "/tasks/{id}": {
      "get": {
        "summary": "Get a task by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Task found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoTaskResponse" }
              }
            }
          },
          "404": { "description": "Task not found" }
        }
      },
      "patch": {
        "summary": "Update a task partially",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TodoTaskUpdate" }
            }
          }
        },
        "responses": {
          "200": { "description": "Task updated" }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "Deleted successfully" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PersonCreate": {
        "type": "object",
        "properties": {
          "firstName": { "type": "string" },
          "lastName": { "type": "string" }
        },
        "required": ["firstName", "lastName"]
      },
      "PersonUpdate": {
        "type": "object",
        "properties": {
          "firstName": { "type": "string" },
          "lastName": { "type": "string" }
        }
      },
      "PersonResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" }
        }
      },

      "TodoItemCreate": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "deadline": { "type": "string", "format": "date" },
          "done": { "type": "boolean" },
          "assigneeId": { "type": "integer" }
        },
        "required": ["title"]
      },
      "TodoItemUpdate": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "deadline": { "type": "string", "format": "date" },
          "done": { "type": "boolean" },
          "assigneeId": { "type": "integer" }
        }
      },
      "TodoItemResponse": {
        "type": "object",
        "properties": {
          "todoId": { "type": "integer" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "deadline": { "type": "string", "format": "date" },
          "done": { "type": "boolean" },
          "assigneeId": { "type": "integer" }
        }
      },

      "TodoTaskCreate": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "done": { "type": "boolean" },
          "todoItemId": { "type": "integer" }
        },
        "required": ["title", "todoItemId"]
      },
      "TodoTaskUpdate": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "done": { "type": "boolean" },
          "todoItemId": { "type": "integer"}
        }
      },
      "TodoTaskResponse": {
        "type": "object",
        "properties": {
          "taskId": { "type": "integer" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "done": { "type": "boolean" },
          "todoItemId": { "type": "integer" }
        }
      }
    }
  }
}